+++ tech1-framework-configurations/src/test/java/io/tech1/framework/configurations/mvc/ApplicationMvc2Test.java
@@ -33,7 +33,7 @@ class ApplicationMvc2Test {
         @Bean
         ApplicationFrameworkProperties applicationFrameworkProperties() {
             var applicationFrameworkProperties = mock(ApplicationFrameworkProperties.class);
-            var mvcConfigs = MvcConfigs.of(
+            var mvcConfigs = new MvcConfigs(
                     true,
                     randomString(),
                     CorsConfigs.of(