+++ tech1-framework-configurations/src/test/java/io/tech1/framework/configurations/mvc/ApplicationMvc3Test.java
@@ -32,7 +32,7 @@ class ApplicationMvc3Test {
         @Bean
         ApplicationFrameworkProperties applicationFrameworkProperties() {
             var applicationFrameworkProperties = mock(ApplicationFrameworkProperties.class);
-            var mvcConfigs = MvcConfigs.of(
+            var mvcConfigs = new MvcConfigs(
                     true,
                     "/framework/security",
                     CorsConfigs.of(