+++ tech1-framework-domain/src/main/java/io/tech1/framework/domain/properties/configs/MvcConfigs.java
@@ -3,10 +3,13 @@ package io.tech1.framework.domain.properties.configs;
 import io.tech1.framework.domain.properties.annotations.MandatoryProperty;
 import io.tech1.framework.domain.properties.annotations.NonMandatoryProperty;
 import io.tech1.framework.domain.properties.configs.mvc.CorsConfigs;
+import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
+import org.springframework.boot.context.properties.ConstructorBinding;
 
 // Lombok (property-based)
+@AllArgsConstructor(onConstructor = @__({@ConstructorBinding}))
 @Data
 @EqualsAndHashCode(callSuper = true)
 public class MvcConfigs extends AbstractPropertiesToggleConfigs {
@@ -16,16 +19,4 @@ public class MvcConfigs extends AbstractPropertiesToggleConfigs {
     private String frameworkBasePathPrefix;
     @NonMandatoryProperty
     private CorsConfigs corsConfigs;
-
-    public static MvcConfigs of(
-            boolean enabled,
-            String frameworkBasePathPrefix,
-            CorsConfigs corsConfigs
-    ) {
-        var instance = new MvcConfigs();
-        instance.enabled = enabled;
-        instance.frameworkBasePathPrefix = frameworkBasePathPrefix;
-        instance.corsConfigs = corsConfigs;
-        return instance;
-    }
 }