+++ tech1-framework-domain/src/test/java/io/tech1/framework/domain/properties/utilties/PropertiesAsserterAndPrinterTest.java
@@ -82,7 +82,7 @@ class PropertiesAsserterAndPrinterTest {
     @Test
     void mvcConfigsDisabledTest() {
         // Arrange
-        var mvcConfigs = MvcConfigs.of(
+        var mvcConfigs = new MvcConfigs(
                 false,
                 null,
                 null